// client/src/types/enums.ts
export enum ThesisType {
  BACHELOR = 'BACHELOR',
  MASTER = 'MASTER',
}

export enum ThesisState {
  PROPOSED = 'PROPOSED',
  IN_PROGRESS = 'IN_PROGRESS',
  SUBMITTED = 'SUBMITTED',
  GRADED = 'GRADED',
  COMPLETED = 'COMPLETED',
}

export enum Language {
  DE = 'DE',
  EN = 'EN',
}

export enum PresentationType {
  PROPOSAL = 'PROPOSAL',
  FINAL = 'FINAL',
}

export enum ApplicationState {
  PENDING = 'PENDING',
  ACCEPTED = 'ACCEPTED',
  REJECTED = 'REJECTED',
  WITHDRAWN = 'WITHDRAWN',
}